requirements.txt
flask>=1.0.2
flask_uploads>=0.2.1
flask_httpauth>=3.2.4
pyyaml>=3.13

